requirements.txt
aiofiles==22.1.0
aiohttp==3.9.2
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==4.2.0
asyncmy==0.2.9
attrs==23.2.0
cffi==1.16.0
choicelib==0.1.5
cryptography==42.0.2
frozenlist==1.4.1
idna==3.6
iso8601==1.1.0
loguru==0.7.2
multidict==6.0.4
pycparser==2.21
pydantic==1.10.14
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2023.4
sniffio==1.3.0
tortoise-orm==0.20.0
typing_extensions==4.9.0
vbml==1.1.post1
vkbottle==4.3.12
vkbottle-types==5.131.146.16
watchfiles==0.18.1
yarl==1.9.4
